@@ -6,7 +6,7 @@ publish_to: "none"
 version: 3.12.0-final+30
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
