@@ -6,7 +6,7 @@ import 'package:share_plus/share_plus.dart';
 import '../../utils/url_launcher.dart';
 
 class Donate extends StatelessWidget {
-  Donate({Key key}) : super(key: key);
+  Donate({Key? key}) : super(key: key);
 
   final _links = {
     'buymeacoffee': 'https://www.buymeacoffee.com/iqfareez',
@@ -15,7 +15,7 @@ class Donate extends StatelessWidget {
     'getapp': 'https://flutree.page.link/getapp'
   };
 
-  void copyToClipboard(String link) {
+  void copyToClipboard(String? link) {
     Clipboard.setData(ClipboardData(text: link))
         .then((value) => Fluttertoast.showToast(msg: 'Copied link'));
   }
@@ -48,14 +48,14 @@ class Donate extends StatelessWidget {
                 DonateCard(
                   label: 'Paypal',
                   link: _links['paypal'],
-                  onTapFun: () => launchURL(context, _links['paypal']),
+                  onTapFun: () => launchURL(context, _links['paypal']!),
                   onLongTapFun: () => copyToClipboard(_links['paypal']),
                   faIcon: FontAwesomeIcons.paypal,
                 ),
                 DonateCard(
                   link: _links['patreon'],
                   faIcon: FontAwesomeIcons.patreon,
-                  onTapFun: () => launchURL(context, _links['patreon']),
+                  onTapFun: () => launchURL(context, _links['patreon']!),
                   onLongTapFun: () => copyToClipboard(_links['patreon']),
                   label: 'Patreon',
                 ),
@@ -63,7 +63,7 @@ class Donate extends StatelessWidget {
                     label: 'Buy me a coffee',
                     link: _links['buymeacoffee'],
                     faIcon: FontAwesomeIcons.mugHot,
-                    onTapFun: () => launchURL(context, _links['buymeacoffee']),
+                    onTapFun: () => launchURL(context, _links['buymeacoffee']!),
                     onLongTapFun: () =>
                         copyToClipboard(_links['buymeacoffee'])),
                 const Text(
@@ -89,12 +89,12 @@ class Donate extends StatelessWidget {
 
 class DonateCard extends StatelessWidget {
   const DonateCard({
-    Key key,
-    @required String label,
-    @required String link,
-    @required IconData faIcon,
-    @required VoidCallback onTapFun,
-    @required VoidCallback onLongTapFun,
+    Key? key,
+    required String label,
+    required String? link,
+    required IconData faIcon,
+    required VoidCallback onTapFun,
+    required VoidCallback onLongTapFun,
   })  : _link = link,
         _label = label,
         _faIcon = faIcon,
@@ -102,7 +102,7 @@ class DonateCard extends StatelessWidget {
         _onLongTapFun = onLongTapFun,
         super(key: key);
 
-  final String _link;
+  final String? _link;
   final String _label;
   final IconData _faIcon;
   final VoidCallback _onTapFun;
@@ -117,7 +117,7 @@ class DonateCard extends StatelessWidget {
         child: ListTile(
           leading: FaIcon(_faIcon),
           title: Text(_label),
-          subtitle: Text(_link),
+          subtitle: Text(_link!),
         ),
       ),
     );
